Separates the internal representation of filter selection and the external api for selection data source. This should remove a lot of unnecessary calls to ParameterBasedFilterInfoSelectionDataSource delegate, there should max be 1 call for each function call in FilterSelectionDataSource-implementation
